+++ CHANGELOG.md
@@ -6,3 +6,10 @@
 * Added missing providers/chains.
 * Added handling of HTTP status code.
 * StandardRelayerApp now can use custom sourceTx options.
+
+## 0.3.4 (2024-04-19)
+
+* Added support for WormholeScan API Key.
+* Redis-storage honours failed queue size.
+* Wallet middleware supports Sepolia.
+